refactor the theme structure
