+++ src/mailingui/utils/theme.ts
@@ -1,32 +1,8 @@
 import { createStitches } from "@stitches/react";
 
 /**
- * This instance of Stitches is for managing token themes
+ * This file is providing utilities for using themes in emails. As a theme, we can use a plain object.
  */
-const theme = {
-  bg: {
-    default: "#f9fafb",
-    subtle: "#F1F5F9",
-    primary: "#2563EB",
-    secondary: "#fff",
-    success: "#f0fdf4",
-    warning: "#fffbeb",
-    danger: "#fef2f2",
-  },
-  text: {
-    default: "#4b5563",
-    subtle: "#2563EB",
-    primary: "#fff",
-    secondary: "#1E293B",
-    success: "#059669",
-    warning: "#b45309",
-    danger: "#b91c1c",
-  },
-  rounded: {
-    md: "4px",
-    full: "9999px",
-  },
-} as const;
 
 /**
  * This instance of Stitches is for styling components and injecting styles into emails
@@ -53,4 +29,4 @@ const extractCssText = (input: string): string => {
  */
 const getCssText = () => extractCssText(stitchesGetCssText());
 
-export { css, getCssText, theme };
+export { css, getCssText };
